rename functions and files and refactor code
