+++ src/utils/highlighting/highlightSearchedFulltexts.js
