+++ src/utils/highlighting/highlightSearchedFulltexts.scss
